+++ src/bindings/capi/thorvg_capi.h
@@ -2173,7 +2173,7 @@ TVG_API Tvg_Animation* tvg_animation_new();
 *
 * \see tvg_animation_get_total_frame()
 */
-TVG_API Tvg_Result tvg_animation_set_frame(Tvg_Animation* animation, uint32_t no);
+TVG_API Tvg_Result tvg_animation_set_frame(Tvg_Animation* animation, float no);
 
 
 /*!
@@ -2205,7 +2205,7 @@ TVG_API Tvg_Paint* tvg_animation_get_picture(Tvg_Animation* animation);
 * \see tvg_animation_get_total_frame()
 * \see tvg_animation_set_frame()
 */
-TVG_API Tvg_Result tvg_animation_get_frame(Tvg_Animation* animation, uint32_t* no);
+TVG_API Tvg_Result tvg_animation_get_frame(Tvg_Animation* animation, float* no);
 
 
 /*!
@@ -2221,7 +2221,7 @@ TVG_API Tvg_Result tvg_animation_get_frame(Tvg_Animation* animation, uint32_t* n
 * \note Frame numbering starts from 0.
 * \note If the Picture is not properly configured, this function will return 0.
 */
-TVG_API Tvg_Result tvg_animation_get_total_frame(Tvg_Animation* animation, uint32_t* cnt);
+TVG_API Tvg_Result tvg_animation_get_total_frame(Tvg_Animation* animation, float* cnt);
 
 
 /*!
